requirements.txt
streamlit
PyPDF2
python-docx
openai
langchain
langchain-community
faiss-cpu
whisper
torch
pytesseract
Pillow
python-dotenv
numpy
tiktoken